Allow Compaction class to potentially handle more than two levels at a time.
Summary:
This diff includes an internal structural change Compaction class, which allows
it to potentially handle more than two levels at a time.  It is the first step
of enabling level compaction to compact multiple levels at a time, and there
is no behavioral changes in this diff.  Some TODOs are added for the following
steps of multi-level compaction.

Test Plan: make all check passed

Reviewers: haobo, ljin, igor, sdong

Subscribers: leveldb

Differential Revision: https://reviews.facebook.net/D19263